fix: fix for styled-components
