+++ packages/preact-timeline/config/noenv.js
@@ -1,0 +1,5 @@
+module.exports = {
+    intro: function intro() {
+        return `var process = { env: {NODE_ENV: "production"} };`;
+    },
+};
