+++ packages/preact-timeline/rollup.config.js
@@ -1,26 +1,10 @@
 import builder from "@daybrush/builder";
-import cjs from "rollup-plugin-commonjs";
-import resolve from "rollup-plugin-node-resolve";
+import css from "rollup-plugin-css-bundle";
+const resolve = require("./config/resolve");
+const commonjs = require("./config/commonjs");
+const noenv = require("./config/noenv");
 
 
-const cjsPlugin = cjs({
-    namedExports: {
-        "node_modules/preact-compat/dist/preact-compat.min.js": ["findDOMNode"],
-    }
-});
-
-const resolvePlugin = resolve();
-const customResolvePlugin =  {
-    ...resolvePlugin,
-    resolveId(importee, importer) {
-        if (importee === "react" || importee === "react-dom") {
-            return resolvePlugin.resolveId("preact-compat", importer);
-        } else if (importee === "prop-types") {
-            return resolvePlugin.resolveId(process.cwd() + "/src/preact-timeline/no-prop-types.ts", importer);
-        }
-        return resolvePlugin.resolveId(importee, importer);
-    },
-}
 
 const defaultOptions = {
     tsconfig: "tsconfig.build.json",
@@ -32,8 +16,9 @@ const defaultOptions = {
         "preact": "Preact",
     },
     plugins: [
-        customResolvePlugin,
-        cjsPlugin,
+        resolve,
+        commonjs,
+        noenv,
     ],
 };
