+++ packages/preact-timeline/src/compat/no-prop-types.js
