+++ packages/preact-timeline/src/compat/styled-components.js
@@ -1,0 +1,3 @@
+import { StyleSheetConsumer } from "styled-components";
+
+StyleSheetConsumer.prototype.shouldComponentUpdate = () => true;
